added tests for units and some for itergration. Added 10000 bugs

+++ backend/controllers/todoListController.js
@@ -15,6 +15,7 @@ module.exports = {
         const userID = req.user.userId
         let todoLists = {}
         // get specific
+        
         if(req.query.todoListId) {
             todoLists = await todoListModel.getTodoList(req.query.todoListId)
         }
@@ -24,17 +25,26 @@ module.exports = {
         }
         // get one or many
         else {
-            console.log('get this lad his lists')
+            // console.log('get this lad his lists')
             todoLists = await todoListModel.getTodoLists(userID)
         }
         let status = todoLists ? 200 : 500;
         res.status(status).json({todoLists});
     },
-    updateTodoList: async (req, res) => {
+    updateTodoList: async (req, res) => { 
+
+        const listOwner = await todoListModel.getTodoList({_id: req.params.todoListId})
         const list = {
             listId : req.params.todoListId
         }
+
+        if(!req.user.isListCollaborator(listOwner.userIds)) {
+            // console.log('incorrect user is making the request')
+            return res.status(401).json({errormsg: 'incorrect user is trying to edit this user'})
+        }
         if(req.body.title) list.title = req.body.title
+
+        // Checks if we should add a user
         if(req.query.userToAdd) list.userId = req.query.userToAdd
 
         let updatedId = await todoListModel.updateTodoList(list)
@@ -42,7 +52,12 @@ module.exports = {
         res.status(status).json({updatedId: updatedId});
     },
     deleteTodoList: async (req, res) => {
-        console.log('time to delete bro' + req.params.todoListId )
+        // console.log('time to delete bro' + req.params.todoListId )
+        const listOwner = await todoListModel.getTodoList({_id: req.params.todoListId})
+        if(!req.user.isListCollaborator(listOwner.userIds)) {
+            // console.log('incorrect user is making the request')
+            return res.status(401).json({errormsg: 'incorrect user is trying to edit this user'})
+        }
         let count = await todoListModel.deleteTodoList(req.params.todoListId)
         let status = count ? 201 : 500;
         res.status(status).json({removed_count: count});
